feat(training): add crash-safe step-based checkpointing and automatic resume
Primary change: step-based, crash-safe checkpointing (ADR 0013).
- checkpoint_every_n_steps / checkpoint_every_minutes: complete atomic bundles
  (model + optimizer + scheduler + RNG + history + step/epoch/metadata) written
  temp -> fsync -> os.replace, to checkpoints/<run_id>/{latest,step_<N>}/.
- --resume auto-locates latest/ and continues at the exact interrupted step
  (deterministic mid-epoch resume; bit-identical to uninterrupted training).
- Ctrl+C saves a final checkpoint and exits cleanly; per-step ETA logging.
- New aion/training/step_checkpoint.py + 9 recovery-validation tests.
- Backward compatible: epoch-based checkpointing remains the default.

Also includes earlier uncommitted AION-0.1 training fixes:
- tokenizer: lossless byte-level BPE pre-tokenization (preserve case + spaces).
- nn: default compute dtype float32 (halves training memory) with float64
  opt-in for gradient-check tests.
- autograd: release graph incrementally in backward() to bound peak memory.
- training pipeline: detailed timed stage/progress logging.

Full suite: 749 passed.

+++ aion/gpt/checkpoint.py
@@ -115,7 +115,7 @@ class GPTCheckpoint:
         for i, p in enumerate(model.parameters()):
             key = param_keys[i] if i < len(param_keys) else _param_key(i, p)
             if key in weights:
-                p.data = weights[key].astype(np.float64)
+                p.data = weights[key].astype(p.data.dtype)
 
         if optimizer is not None and meta.get("optimizer") is not None:
             optim_path = self._optim_path(epoch)
